feat(alembic): add auto-generated migration for source model updates
- Generated new migration f79840d8a11e using alembic revision --autogenerate
- Manually added EntityType 'Source' and SourceType entries to type_lookup table
- Includes field changes: content, source_type_id, source_metadata, source_id in test table
- Replaces manual bffb5364a019 migration with auto-generated equivalent
