+++ apps/backend/src/rhesis/backend/alembic/versions/f79840d8a11e_update_source_model_with_comments_.py
@@ -1,0 +1,85 @@
+from typing import Sequence, Union
+
+import sqlalchemy as sa
+from alembic import op
+from sqlalchemy.dialects import postgresql
+
+import rhesis
+from rhesis.backend.alembic.utils.template_loader import load_type_lookup_template
+
+# revision identifiers, used by Alembic.
+revision: str = "f79840d8a11e"
+down_revision: Union[str, None] = "4df7965dfdfb"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+
+    # Step 1: Add EntityType "Source" to type_lookup table
+    entity_type_values = "('EntityType', 'Source', 'Entity type for sources')"
+    op.execute(load_type_lookup_template(entity_type_values))
+
+    # Step 2: Add SourceType entries to type_lookup table
+    source_type_values = """
+        ('SourceType', 'Document', 'Document source (PDF, Word, etc.)'),
+        ('SourceType', 'Website', 'Website or web page source'),
+        ('SourceType', 'API', 'API documentation or response'),
+        ('SourceType', 'Database', 'Database schema or data source'),
+        ('SourceType', 'Code', 'Source code or repository'),
+        ('SourceType', 'Manual', 'Manual or user-provided content')
+    """.strip()
+
+    # ### commands auto generated by Alembic - please adjust! ###
+
+    op.execute(load_type_lookup_template(source_type_values))
+
+    op.alter_column("demographic", "name", existing_type=sa.VARCHAR(), nullable=False)
+    op.alter_column("dimension", "name", existing_type=sa.VARCHAR(), nullable=False)
+    op.add_column("source", sa.Column("content", sa.Text(), nullable=True))
+    op.add_column(
+        "source", sa.Column("source_type_id", rhesis.backend.app.models.guid.GUID(), nullable=True)
+    )
+    op.add_column(
+        "source",
+        sa.Column("source_metadata", postgresql.JSONB(astext_type=sa.Text()), nullable=True),
+    )
+    op.create_foreign_key(None, "source", "type_lookup", ["source_type_id"], ["id"])
+    op.drop_column("source", "entity_type")
+    op.add_column(
+        "test", sa.Column("source_id", rhesis.backend.app.models.guid.GUID(), nullable=True)
+    )
+    op.create_foreign_key(None, "test", "source", ["source_id"], ["id"])
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint(None, "test", type_="foreignkey")
+    op.drop_column("test", "source_id")
+    op.add_column(
+        "source", sa.Column("entity_type", sa.VARCHAR(), autoincrement=False, nullable=True)
+    )
+    op.drop_constraint(None, "source", type_="foreignkey")
+    op.drop_column("source", "source_metadata")
+    op.drop_column("source", "source_type_id")
+    op.drop_column("source", "content")
+    op.alter_column("dimension", "name", existing_type=sa.VARCHAR(), nullable=True)
+    op.alter_column("demographic", "name", existing_type=sa.VARCHAR(), nullable=True)
+
+    # ### end Alembic commands ###
+
+    # Step 1: Remove SourceType entries from type_lookup table
+    source_type_values = "'Document', 'Website', 'API', 'Database', 'Code', 'Manual'"
+    op.execute(f"""
+        DELETE FROM type_lookup
+        WHERE type_name = 'SourceType'
+        AND type_value IN ({source_type_values})
+    """)
+
+    # Step 2: Remove EntityType "Source" from type_lookup table
+    op.execute("""
+        DELETE FROM type_lookup
+        WHERE type_name = 'EntityType'
+        AND type_value = 'Source'
+    """)
